use raw genotype likelihoods instead of priors

+++ src/phredgenotypelikelihoods.h
@@ -5,14 +5,14 @@
 
 class PhredGenotypeLikelihoods {
 public:
-	PhredGenotypeLikelihoods(unsigned int gl0 = 0, unsigned int gl1 = 0, unsigned int gl2 = 0);
+	PhredGenotypeLikelihoods(double gl0 = 0, double gl1 = 0, double gl2 = 0);
 
-	unsigned int get(size_t genotype) const;
+	double get(size_t genotype) const;
 
 	std::string toString() const;
 
 private:
-	std::array<unsigned int, 3> gl;
+	std::array<double, 3> gl;
 };
