+++ whatshap/core.pyx
@@ -335,7 +335,7 @@ cdef class Pedigree:
 
 
 cdef class PhredGenotypeLikelihoods:
-	def __cinit__(self, int gl0 = 0, int gl1 = 0, int gl2 = 0):
+	def __cinit__(self, double gl0 = 0, double gl1 = 0, double gl2 = 0):
 		self.thisptr = new cpp.PhredGenotypeLikelihoods(gl0, gl1, gl2)
 
 	def __dealloc__(self):
